moving dataplane client from pkg to dataplaneapi package
