+++ go.mod
@@ -10,7 +10,7 @@ require (
 	github.com/spf13/cobra v1.6.1
 	github.com/spf13/pflag v1.0.5
 	github.com/spf13/viper v1.14.0
-	github.com/stretchr/testify v1.8.1
+	github.com/stretchr/testify v1.8.2
 	go.infratographer.com/x v0.0.3
 	go.uber.org/zap v1.24.0
 	gocloud.dev v0.27.0
