+++ internal/dataplaneapi/errors.go
@@ -1,4 +1,4 @@
-package pkg
+package dataplaneapi
 
 import "errors"
